Fix more golangci-lint issues
