+++ pkg/benchmark/compression.go
@@ -37,26 +37,32 @@ func Lz4() CompressionAlgorithm {
 func (c *Lz4CompressionAlgo) Compress(data []byte) ([]byte, error) {
 	buf := make([]byte, len(data))
 	ht := make([]int, 64<<10)
+
 	n, err := lz4.CompressBlock(data, buf, ht)
 	if err != nil {
 		return nil, err
 	}
+
 	if n == 0 {
 		// uncompressible
 		buf = data
 	} else {
 		buf = buf[:n]
 	}
+
 	return buf, nil
 }
 
 func (c *Lz4CompressionAlgo) Decompress(data []byte) ([]byte, error) {
 	decompressed := make([]byte, 10*len(data))
+
 	n, err := lz4.UncompressBlock(data, decompressed)
 	if err != nil {
 		return nil, err
 	}
+
 	decompressed = decompressed[:n]
+
 	return decompressed, nil
 }
 
@@ -74,6 +80,7 @@ func Zstd() CompressionAlgorithm {
 	if err != nil {
 		panic(err)
 	}
+
 	decoder, err := zstd.NewReader(nil)
 	if err != nil {
 		panic(err)
@@ -85,7 +92,9 @@ func Zstd() CompressionAlgorithm {
 func (c *ZstdCompressionAlgo) Compress(data []byte) ([]byte, error) {
 	var buf bytes.Buffer
 	c.encoder.Reset(&buf)
+
 	compressed := c.encoder.EncodeAll(data, nil)
+
 	return compressed, nil
 }
 
@@ -94,6 +103,7 @@ func (c *ZstdCompressionAlgo) Decompress(data []byte) ([]byte, error) {
 	if err != nil {
 		return nil, err
 	}
+
 	decompressed, err := c.decoder.DecodeAll(data, nil)
 	if err != nil {
 		return nil, err
