+++ pkg/benchmark/profileable/arrow/otlp_arrow_profiler_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package otlp_arrow
+package arrow
 
 import (
 	"path/filepath"
