+++ pkg/otel/traces/validation_test.go
@@ -93,6 +93,8 @@ func TestConversionFromRealData(t *testing.T) {
 }
 
 func checkTracesConversion(t *testing.T, expectedRequest ptraceotlp.ExportRequest) { //nolint:unused // only used for testing
+	t.Helper()
+
 	pool := memory.NewCheckedAllocator(memory.NewGoAllocator())
 	defer pool.AssertSize(t, 0)
 	traceSchema := acommon.NewAdaptiveSchema(tracesarrow.Schema)
